requirements.txt
opencv-python>=4.5.3
numpy>=1.21.2
Pillow>=8.3.1
matplotlib>=3.4.3
torch>=1.9.1
torchvision>=0.10.0
tensorboard>=2.6.0
pandas>=1.3.2
scikit-learn>=0.24.2
tqdm>=4.62.2
googledrivedownloader>=0.4
recap>=0.1.6
chess>=1.6.1
osfclient>=0.0.5
setuptools==59.5.0
paho-mqtt==1.6.1
python-dotenv==1.0.0
